Updates tests to use Floss/Electron
