+++ package.json
@@ -21,8 +21,9 @@
     "start": "parallelshell \"npm run watch:debug\" \"npm run watch:release\"",
     "watch:debug": "watchify src/index.js -s PIXI -o bin/pixi.js -dv",
     "watch:release": "watchify src/index.js -s PIXI -o bin/pixi.min.js -dv",
-    "test": "npm run build && testem ci",
-    "lint": "jshint --reporter=scripts/reporter.js scripts src",
+    "test": "floss --path test/index.js",
+    "test:debug": "npm test -- --debug",
+    "lint": "jshint --reporter=scripts/reporter.js scripts src test",
     "prebuild": "npm run lint && npm run clean",
     "build": "node scripts/bundle.js",
     "docs": "jsdoc -c scripts/jsdoc.conf.json -R README.md"
@@ -49,8 +50,9 @@
   "devDependencies": {
     "browserify": "^13.0.1",
     "browserify-versionify": "^1.0.6",
-    "chai": "^3.5.0",
     "del": "^2.2.0",
+    "electron-prebuilt": "^1.3.2",
+    "floss": "^0.7.1",
     "gulp-empty": "^0.1.1",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^2.0.0",
@@ -61,10 +63,8 @@
     "jshint-stylish-summary": "^0.1.2",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "mocha": "^2.5.3",
     "parallelshell": "^2.0.0",
     "rimraf": "^2.5.3",
-    "testem": "^1.8.1",
     "through2": "^2.0.1",
     "vinyl-buffer": "^1.0.0",
     "vinyl-fs": "^2.4.3",
