+++ test/.jshintrc
@@ -1,4 +1,13 @@
 {
-    "mocha": true,
-    "chai": true
+  "node": true,
+  "mocha": true,
+  "browser": true,
+  "esversion": 6,
+  "expr": true,
+  "predef": [
+  	"PIXI",
+  	"chai",
+  	"expect",
+    "assert"
+  ]
 }
