+++ test/core/Text.js
