+++ test/core/index.js
@@ -1,0 +1,9 @@
+require('./Bounds');
+require('./Container');
+require('./DisplayObject');
+require('./getLocalBounds');
+require('./Sprite');
+require('./Text');
+require('./toGlobal');
+require('./toLocal');
+require('./util');
