+++ test/core/toLocal.js
@@ -1,6 +1,5 @@
 describe('toLocal', function () {
 
-
     it('should return correct local cordinates of a displayObject', function() {
 
         var parent = new PIXI.Container();
@@ -11,7 +10,9 @@ describe('toLocal', function () {
 
         var point = new PIXI.Point(100, 100);
 
-        var localPoint = container.toLocal(point);
+        var localPoint;
+
+        localPoint = container.toLocal(point);
 
         expect(localPoint.x).to.equal(100);
         expect(localPoint.y).to.equal(100);
@@ -22,7 +23,7 @@ describe('toLocal', function () {
         container.scale.x = 2;
         container.scale.y = 2;
 
-        var localPoint = container.toLocal(point);
+        localPoint = container.toLocal(point);
 
         expect(localPoint.x).to.equal(40);
         expect(localPoint.y).to.equal(40);
@@ -43,9 +44,6 @@ describe('toLocal', function () {
         container2.position.y = 100;
 
         var point = new PIXI.Point(100, 100);
-
-       // container.position.x = 20;
-       // container.position.y = 20;
 
         container.scale.x = 2;
         container.scale.y = 2;
