+++ test/index.js
@@ -1,0 +1,9 @@
+require('../bin/pixi');
+
+describe('PIXI', function () {
+    it('should exist as a global object', function () {
+        expect(PIXI).to.be.an('object');
+    });
+    require('./core');
+    require('./interaction');
+});
