--- test/setup.js
@@ -1,1 +1,0 @@
-expect = chai.expect;
