--- test/unit/index.test.js
@@ -1,6 +1,0 @@
-describe('PIXI', function () {
-    it('exists', function () {
-        expect(PIXI)
-            .to.be.an('object');
-    });
-});
