--- testem.json
@@ -1,17 +1,0 @@
-{
-    "framework" : "mocha+chai",
-    "launch_in_dev" : ["Chrome", "Firefox"],
-    "launch_in_ci" : ["Firefox"],
-    "src_files" : [
-        "bin/pixi.js",
-
-        "test/setup.js",
-        "test/unit/**/*.test.js"
-    ],
-    "launchers": {
-        "node": {
-            "command": "./node_modules/.bin/mocha -r ./test/setup.js -R tap --recursive ./test/unit/",
-            "protocol": "tap"
-        }
-    }
-}
